Add benchmarks for JSON deserialization including randomized map keys
Most benchmarks use the same set of input data for each iteration,
which is heavily biased toward caches. This benchmark is meant to
stress canonicalzation cache misses in a way that we observe in
some production systems.

My implementation isn't ideal because I'm doing a fair bit of work
to generate randomized data within the measured component of the
benchmark, however I've profiled the results and most time is
not spent within the setup portion, and the benchmark setup is
comparable between configurations such that the results should
be comparable between flags.

Initially I began this investigation based on the InternCache,
which I expected to be the primary bottle-neck for reasons
listed here: https://shipilev.net/jvm/anatomy-quarks/10-string-intern/
There is a measurable impact disabling interning, particularly
with smaller heap sizes, but there's a much larger improvement
when we opt out of canonicalization entirely, especially in
paths which rely on the ByteQuadsCanonicalizer rather than
CharsToNameCanonicalizer.